wallet-guardian/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
spoon-ai-sdk>=0.3.4
pytest>=8.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0

